+++ OpenTaiko/src/Stages/07.Game/Taiko/CAct演奏DrumsRunner.cs
@@ -64,52 +64,7 @@ namespace TJAPlayer3
                 stRunners[i].ct進行 = new CCounter();
             }
             
-            string presetSection = "";
-            if (TJAPlayer3.stage選曲.n確定された曲の難易度[0] == (int)Difficulty.Tower)
-            {
-                presetSection = "Tower";
-            }
-            else if (TJAPlayer3.stage選曲.n確定された曲の難易度[0] == (int)Difficulty.Dan)
-            {
-                presetSection = "Dan";
-            }
-            else if (TJAPlayer3.ConfigIni.bAIBattleMode)
-            {
-            }
-            else
-            {
-                presetSection = "Regular";
-            }
-
-            object _ps = null;
-
-            switch (presetSection)
-            {
-                case "Regular":
-                    _ps = TJAPlayer3.Skin.Game_SkinScenes.Regular;
-                    break;
-                case "Dan":
-                    _ps = TJAPlayer3.Skin.Game_SkinScenes.Dan;
-                    break;
-                case "Tower":
-                    _ps = TJAPlayer3.Skin.Game_SkinScenes.Tower;
-                    break;
-                default:
-                    break;
-            };
-            
-            var preset = (_ps != null 
-                    && TJAPlayer3.stage選曲.r確定された曲.strScenePreset != null 
-                    && ((Dictionary<string, DBSkinPreset.SkinScene>)_ps).ContainsKey(TJAPlayer3.stage選曲.r確定された曲.strScenePreset)) 
-                ? ((Dictionary<string,DBSkinPreset.SkinScene>)_ps)[TJAPlayer3.stage選曲.r確定された曲.strScenePreset] 
-                : null;
-
-            if (_ps != null
-                    && TJAPlayer3.DTX.scenePreset != null
-                    && ((Dictionary<string, DBSkinPreset.SkinScene>)_ps).ContainsKey(TJAPlayer3.DTX.scenePreset)) // If currently selected song has valid SCENEPRESET metadata within TJA
-            {
-                preset = ((Dictionary<string, DBSkinPreset.SkinScene>)_ps)[TJAPlayer3.DTX.scenePreset];
-            }
+            var preset = HScenePreset.GetBGPreset();
             
             Random random = new Random();
 
@@ -119,7 +74,7 @@ namespace TJAPlayer3
                 var dirs = System.IO.Directory.GetDirectories($@"{dancerOrigindir}");
                 if (dirs.Length > 0)
                 {
-                    var _presetPath = (preset != null) ? $@"{dancerOrigindir}" + preset.RunnerSet[random.Next(0, preset.RunnerSet.Length)] : "";
+                    var _presetPath = (preset != null && preset.RunnerSet != null) ? $@"{dancerOrigindir}" + preset.RunnerSet[random.Next(0, preset.RunnerSet.Length)] : "";
                     var path = (preset != null && System.IO.Directory.Exists(_presetPath)) 
                         ?  _presetPath
                         : dirs[random.Next(0, dirs.Length)];
